requirements.txt
torch
torchvision
pillow
torchnet
tqdm
keras==2.1.2
opencv-python
scikit-learn
h5py
tensorflow==1.11
pandas
matplotlib